+++ example/app/kendo/kendo-example.component.html
@@ -8,13 +8,13 @@
                                     [model]="controlModel"
                                     (change)="onChange($event)">
 
-            <ng-template modelId="kendoDropDownList" type="kendoDropDownListFooterTemplate">
+            <ng-template modelId="kendoDropDownList" as="kendoDropDownListFooterTemplate">
 
                 <p>LIST FOOTER TEMPLATE</p>
 
             </ng-template>
 
-            <ng-template modelId="kendoDropDownList" type="kendoDropDownListHeaderTemplate">
+            <ng-template modelId="kendoDropDownList" as="kendoDropDownListHeaderTemplate">
 
                 <p>LIST HEADER TEMPLATE</p>
