+++ modules/ui-material/src/dynamic-form-material.const.ts
@@ -1,14 +1,3 @@
-import {
-    MdAutocomplete,
-    MdCheckbox,
-    MdDatepicker,
-    MdInputContainer,
-    MdRadioGroup,
-    MdSelect,
-    MdSlider,
-    MdSlideToggle
-} from "@angular/material";
-
 export const enum MdFormControlType {
 
     Array = 1, //"ARRAY",
@@ -22,9 +11,6 @@ export const enum MdFormControlType {
     SlideToggle = 9, //"SLIDE_TOGGLE",
     TextArea = 10, //"TEXTAREA"
 }
-
-export type MdFormControlComponent = MdAutocomplete | MdCheckbox | MdDatepicker<Date> | MdInputContainer |
-    MdRadioGroup | MdSelect | MdSlider | MdSlideToggle;
 
 export const MD_VIEW_CHILD_SELECTOR = `mdAutocomplete,mdCheckbox,mdDatepicker,mdInput,mdRadioGroup,mdSelect,
     mdSlider,mdSlideToggle,mdTextarea`;
